requirements.txt
chonkie
torch
sentence-transformers
python-dotenv
argparse
loguru
